session: move some test into new package
