+++ pkg/session/BUILD.bazel
@@ -127,7 +127,6 @@ go_test(
         "main_test.go",
         "tidb_test.go",
     ],
-    data = glob(["testdata/**"]),
     embed = [":session"],
     flaky = True,
     race = "on",
@@ -143,12 +142,10 @@ go_test(
         "//pkg/meta",
         "//pkg/parser/ast",
         "//pkg/parser/auth",
-        "//pkg/planner/core",
         "//pkg/sessionctx",
         "//pkg/sessionctx/variable",
         "//pkg/statistics",
         "//pkg/store/mockstore",
-        "//pkg/tablecodec",
         "//pkg/telemetry",
         "//pkg/testkit/testmain",
         "//pkg/testkit/testsetup",
