+++ pkg/session/test/BUILD.bazel
@@ -6,9 +6,10 @@ go_test(
     srcs = [
         "main_test.go",
         "session_test.go",
+        "tidb_test.go",
     ],
     flaky = True,
-    shard_count = 26,
+    shard_count = 28,
     deps = [
         "//pkg/config",
         "//pkg/domain",
@@ -19,8 +20,10 @@ go_test(
         "//pkg/parser/model",
         "//pkg/parser/mysql",
         "//pkg/parser/terror",
+        "//pkg/planner/core",
         "//pkg/session",
         "//pkg/store/mockstore",
+        "//pkg/tablecodec",
         "//pkg/testkit",
         "//pkg/testkit/testmain",
         "//pkg/testkit/testsetup",
